Puesta en común y corrección de errores.

+++ BD/init.sql
@@ -2,8 +2,9 @@ DROP TABLE cuentas;
 DROP TABLE inventario;
 CREATE TABLE cuentas(
 id INT,
-total REAL(20),
+total REAL,
 PRIMARY KEY (id));
+
 CREATE TABLE inventario(
 id INT,
 nombre VARCHAR(20),
